Separate collection creation button from toolbar
