+++ client/src/components/OverviewToolbar.vue
@@ -1,24 +1,11 @@
 <script setup lang="ts">
-import { computed, ComputedRef, onMounted, ref } from 'vue';
 import { useGuidelinesStore } from '../store/guidelines';
-import DataInputComponent from './DataInputComponent.vue';
-import DataInputGroup from './DataInputGroup.vue';
-import Button from 'primevue/button';
-import Dialog from 'primevue/dialog';
 import IconField from 'primevue/iconfield';
 import InputIcon from 'primevue/inputicon';
 import InputText from 'primevue/inputtext';
 import Tag from 'primevue/tag';
 import MultiSelect from 'primevue/multiselect';
 import Toolbar from 'primevue/toolbar';
-import {
-  buildFetchUrl,
-  capitalize,
-  cloneDeep,
-  getDefaultValueForProperty,
-} from '../utils/helper/helper';
-import ICollection from '../models/ICollection';
-import { CollectionAccessObject, PropertyConfig } from '../models/types';
 
 // Accept the composable's refs as props
 const props = defineProps<{
@@ -28,116 +15,7 @@ const props = defineProps<{
 
 const emit = defineEmits(['collectionCreated', 'searchInputChanged', 'nodeLabelsInputChanged']);
 
-const {
-  availableCollectionLabels,
-  guidelines,
-  getAllCollectionConfigFields,
-  getCollectionConfigFields,
-} = useGuidelinesStore();
-
-const newCollectionData = ref<CollectionAccessObject>(null);
-const dialogIsVisible = ref<boolean>(false);
-const asyncOperationRunning = ref<boolean>(false);
-
-const dialogInputFields: ComputedRef<PropertyConfig[]> = computed(() =>
-  getCollectionConfigFields(newCollectionData.value.collection.nodeLabels),
-);
-
-const inputIsValid: ComputedRef<boolean> = computed((): boolean => {
-  const selectedLabelsAreValid: boolean =
-    availableCollectionLabels.value.length === 0 ||
-    newCollectionData.value.collection.nodeLabels.length > 0;
-
-  const requiredFieldsAreValid: boolean = dialogInputFields.value
-    .filter(field => field.required)
-    .every(
-      field =>
-        newCollectionData.value?.collection.data[field.name] !== null &&
-        newCollectionData.value?.collection.data[field.name] !== undefined &&
-        newCollectionData.value?.collection.data[field.name] !== '',
-    );
-
-  return selectedLabelsAreValid && requiredFieldsAreValid;
-});
-
-function removeUnnecessaryDataBeforeSave(): void {
-  const configuredFieldNames: string[] = getCollectionConfigFields(
-    newCollectionData.value.collection.nodeLabels,
-  ).map(f => f.name);
-
-  Object.keys(newCollectionData.value.collection.data).forEach(key => {
-    if (!configuredFieldNames.includes(key) && key !== 'uuid') {
-      delete newCollectionData.value.collection.data[key];
-    }
-  });
-}
-
-// No longer needed since we sync with props
-onMounted(() => {
-  // Initial emit in case parent needs the default values
-  if (!props.nodeLabelsValue) {
-    emit('nodeLabelsInputChanged', availableCollectionLabels.value);
-  }
-});
-
-async function createNewCollection(): Promise<void> {
-  removeUnnecessaryDataBeforeSave();
-
-  console.log(cloneDeep(newCollectionData.value));
-
-  asyncOperationRunning.value = true;
-
-  try {
-    const url: string = buildFetchUrl('/api/collections');
-
-    const response: Response = await fetch(url, {
-      method: 'POST',
-      cache: 'no-cache',
-      credentials: 'same-origin',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      referrerPolicy: 'no-referrer',
-      body: JSON.stringify(newCollectionData.value),
-    });
-
-    if (!response.ok) {
-      throw new Error('Network response was not ok');
-    }
-
-    const createdCollection: ICollection = await response.json();
-
-    newCollectionData.value = {} as CollectionAccessObject;
-    dialogIsVisible.value = false;
-
-    emit('collectionCreated', createdCollection);
-  } catch (error: unknown) {
-    console.error('Error creating collection:', error);
-  } finally {
-    asyncOperationRunning.value = false;
-  }
-}
-
-async function showDialog(): Promise<void> {
-  dialogIsVisible.value = true;
-
-  newCollectionData.value = {
-    collection: {
-      data: Object.fromEntries(
-        getAllCollectionConfigFields().map(f => [
-          f.name,
-          f?.required === true ? getDefaultValueForProperty(f.type) : null,
-        ]),
-      ) as ICollection,
-      nodeLabels: guidelines.value.collections.types
-        .filter(t => t.level === 'primary')
-        .map(t => t.additionalLabel),
-    },
-    texts: [],
-    annotations: [],
-    collections: {},
-  } as CollectionAccessObject;
-}
+const { availableCollectionLabels } = useGuidelinesStore();
 
 function handleSearchInputChange(value: string) {
   emit('searchInputChanged', value);
@@ -145,11 +23,6 @@ function handleSearchInputChange(value: string) {
 
 function handleNodeLabelsChange(value: string[]) {
   emit('nodeLabelsInputChanged', value);
-}
-
-async function hideDialog(): Promise<void> {
-  newCollectionData.value = null;
-  dialogIsVisible.value = false;
 }
 </script>
 
@@ -189,92 +62,7 @@ async function hideDialog(): Promise<void> {
         />
       </IconField>
     </template>
-
-    <template #end>
-      <Button
-        icon="pi pi-plus"
-        aria-label="Submit"
-        label="Add Collection"
-        title="Add new Collection"
-        @click="showDialog"
-      />
-    </template>
   </Toolbar>
-
-  <Dialog
-    v-model:visible="dialogIsVisible"
-    modal
-    :closable="false"
-    :close-on-escape="false"
-    :style="{ width: '25rem' }"
-  >
-    <template #header>
-      <h2 class="w-full text-center m-0">Add new Collection</h2>
-    </template>
-    <form @submit.prevent="createNewCollection">
-      <div class="select-label-container flex flex-column align-items-center">
-        <h4 class="mt-0">Select labels</h4>
-        <MultiSelect
-          v-model="newCollectionData.collection.nodeLabels"
-          :options="availableCollectionLabels"
-          display="chip"
-          :invalid="
-            availableCollectionLabels?.length > 0 &&
-            newCollectionData?.collection?.nodeLabels?.length === 0
-          "
-          placeholder="Select labels"
-          class="text-center"
-          :filter="false"
-        >
-          <template #chip="{ value }">
-            <Tag :value="value" severity="contrast" class="mr-1" />
-          </template>
-        </MultiSelect>
-      </div>
-      <div class="select-data-container">
-        <h4 class="text-center">Add data</h4>
-        <div
-          class="input-container flex align-items-center gap-3 mb-3"
-          v-for="field in dialogInputFields"
-        >
-          <label :for="field.name" class="font-semibold w-6rem"
-            >{{ capitalize(field.name) }}
-          </label>
-          <DataInputGroup
-            v-if="field.type === 'array'"
-            v-model="newCollectionData.collection.data[field.name]"
-            :config="field"
-          />
-          <DataInputComponent
-            v-else
-            v-model="newCollectionData.collection.data[field.name]"
-            :config="field"
-          />
-        </div>
-      </div>
-
-      <div class="button-container flex justify-content-end gap-2">
-        <Button
-          type="button"
-          label="Cancel"
-          title="Cancel"
-          severity="secondary"
-          @click="hideDialog"
-        ></Button>
-        <Button
-          type="submit"
-          label="Create"
-          title="Create new text"
-          :loading="asyncOperationRunning"
-          :disabled="!inputIsValid"
-        ></Button>
-      </div>
-    </form>
-  </Dialog>
 </template>
 
-<style scoped>
-label {
-  word-wrap: break-word;
-}
-</style>
+<style scoped></style>
